rreplace some subcomponents
